Revert "Soporte inicial para FACeB2B"
This reverts commit 9587745b191fe188ce4acb8a5babd61d282b185b.
