+++ src/Face/FaceClient.php
@@ -3,35 +3,27 @@ namespace josemmo\Facturae\Face;
 
 use josemmo\Facturae\Common\XmlTools;
 
-/**
- * This class allows communication with the FACe Web Service.
- */
-class FaceClient extends FaceAbstractClient {
+class FaceClient extends SoapClient {
+
+  private static $PROD_URL = "https://webservice.face.gob.es/facturasspp2";
+  private static $STAGING_URL = "https://se-face-webservice.redsara.es/facturasspp2";
+
 
   /**
-   * Get production endpoint
-   * @return string Production endpoint
+   * Get endpoint URL
+   * @return string Endpoint URL
    */
-  protected function getProductionEndpoint() {
-    return "https://webservice.face.gob.es/";
+  protected function getEndpointUrl() {
+    return $this->production ? self::$PROD_URL : self::$STAGING_URL;
   }
 
 
   /**
-   * Get staging endpoint
-   * @return string Staging endpoint
+   * Get invoice status codes
+   * @return SimpleXMLElement Response
    */
-  protected function getStagingEndpoint() {
-    return "https://se-face-webservice.redsara.es/";
-  }
-
-
-  /**
-   * Get web namespace
-   * @return string Web namespace
-   */
-  protected function getWebNamespace() {
-    return "https://webservice.face.gob.es";
+  public function getStatus() {
+    return $this->request('<web:consultarEstados></web:consultarEstados>');
   }
 
 
@@ -70,6 +62,68 @@ class FaceClient extends FaceAbstractClient {
     return $this->request('<web:consultarNIFsPorAdministracion>' .
       '<codigoDir>' . $code . '</codigoDir>' .
       '</web:consultarNIFsPorAdministracion>');
+  }
+
+
+  /**
+   * Get invoice
+   * @param  string|string[]  $regId Invoice register ID(s)
+   * @return SimpleXMLElement        Response
+   */
+  public function getInvoices($regId) {
+    if (is_string($regId)) {
+      return $this->request('<web:consultarFactura>' .
+        '<numeroRegistro>' . $regId . '</numeroRegistro>' .
+        '</web:consultarFactura>');
+    }
+    $req = '<web:consultarListadoFacturas><request>';
+    foreach ($regId as $id) $req .= '<numeroRegistro>' . $id . '</numeroRegistro>';
+    $req .= '</request></web:consultarListadoFacturas>';
+    return $this->request($req);
+  }
+
+
+  /**
+   * Send invoice
+   * @param  string           $email       Email address
+   * @param  FacturaeFile     $invoice     Invoice
+   * @param  FacturaeFile[]   $attachments Attachments
+   * @return SimpleXMLElement              Response
+   */
+  public function sendInvoice($email, $invoice, $attachments=array()) {
+    $tools = new XmlTools();
+    $req = '<web:enviarFactura><request>';
+    $req .= '<correo>' . $email . '</correo>';
+    $req .= '<factura>' .
+        '<factura>' . $tools->toBase64($invoice->getData()) . '</factura>' .
+        '<nombre>' . $invoice->getFilename() . '</nombre>' .
+        '<mime>' . $invoice->getMimeType() . '</mime>' .
+      '</factura>';
+    $req .= '<anexos>';
+    foreach ($attachments as $file) {
+      $req .= '<anexo>' .
+          '<anexo>' . $tools->toBase64($file->getData()) . '</anexo>' .
+          '<nombre>' . $file->getFilename() . '</nombre>' .
+          '<mime>' . $file->getMimeType() . '</mime>' .
+        '</anexo>';
+    }
+    $req .= '</anexos>';
+    $req .= '</request></web:enviarFactura>';
+    return $this->request($req);
+  }
+
+
+  /**
+   * Cancel invoice
+   * @param  string           $regId  Invoice register ID
+   * @param  string           $reason Cancelation reason
+   * @return SimpleXMLElement         Response
+   */
+  public function cancelInvoice($regId, $reason) {
+    return $this->request('<web:anularFactura>' .
+      '<numeroRegistro>' . $regId . '</numeroRegistro>' .
+      '<motivo>' . $reason . '</motivo>' .
+      '</web:anularFactura>');
   }
 
 }
